producer-api/requirements.txt
fastapi
uvicorn
pymongo==3.12.1
kafka-python
